backend/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
sqlmodel==0.0.8
asyncpg==0.27.0
httpx==0.24.1
python-dotenv==1.0.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.2
git+https://github.com/sentinel-hub/sentinelhub-py.git
numpy==1.26.0
pandas==2.1.0
psycopg2-binary==2.9.11
python-multipart==0.0.6
